updated cpp files

+++ src/main.cpp
@@ -28,6 +28,7 @@ std::string hasData(std::string s) {
 
 int main()
 {
+
   uWS::Hub h;
 
   // Create a Kalman Filter instance
@@ -60,7 +61,7 @@ int main()
           
           MeasurementPackage meas_package;
           istringstream iss(sensor_measurment);
-    	  long long timestamp;
+    	  int64_t timestamp;
 
     	  // reads first element from the current line
     	  string sensor_type;
@@ -104,7 +105,8 @@ int main()
     	  gt_values(2) = vx_gt;
     	  gt_values(3) = vy_gt;
     	  ground_truth.push_back(gt_values);
-          
+
+				   
           //Call ProcessMeasurment(meas_package) for Kalman filter
     	  fusionEKF.ProcessMeasurement(meas_package);    	  
 
@@ -182,4 +184,7 @@ int main()
     return -1;
   }
   h.run();
+
 }
+
+
